Revise classes to separate concerns
